requirements.txt
fastapi==0.70.1
uvicorn==0.16.0
mysql-connector-python==8.0.27
tweepy==4.4.0
pytest==6.2.5
pytest-cov==3.0.0
python-multipart==0.0.5
jinja2==3.0.3